Port RenderDoc to Qt6 and add ARM64 platform support
- Migrate all Qt5 deprecated APIs to Qt6 equivalents across qrenderdoc
- Update 3rdparty components (ToolWindowManager, Scintilla, FlowLayout)
- Replace qtmain.lib with Qt6EntryPoint.lib
- Add ARM64 platform configurations to all vcxproj files
- Update solution file to remove Win32 and add ARM64
- Update Python from 3.6 to 3.14 (rename natvis, update props)
- Fix rdcstr/rdcarray headers for ARM64 MSVC compatibility
- Update WindowsSDK props for VS 18 Insiders toolchain

+++ qrenderdoc/Widgets/Extended/RDTreeView.h
@@ -145,20 +145,20 @@ public:
   void updateExpansion(RDTreeViewExpansionState &state, int keyColumn, int keyRole = Qt::DisplayRole)
   {
     updateExpansion(state, [keyColumn, keyRole](QModelIndex idx, uint seed) {
-      return qHash(idx.sibling(idx.row(), keyColumn).data(keyRole).toString(), seed);
+      return (uint)qHash(idx.sibling(idx.row(), keyColumn).data(keyRole).toString(), seed);
     });
   }
   void saveExpansion(RDTreeViewExpansionState &state, int keyColumn, int keyRole = Qt::DisplayRole)
   {
     saveExpansion(state, [keyColumn, keyRole](QModelIndex idx, uint seed) {
-      return qHash(idx.sibling(idx.row(), keyColumn).data(keyRole).toString(), seed);
+      return (uint)qHash(idx.sibling(idx.row(), keyColumn).data(keyRole).toString(), seed);
     });
   }
   void applyExpansion(const RDTreeViewExpansionState &state, int keyColumn,
                       int keyRole = Qt::DisplayRole)
   {
     applyExpansion(state, [keyColumn, keyRole](QModelIndex idx, uint seed) {
-      return qHash(idx.sibling(idx.row(), keyColumn).data(keyRole).toString(), seed);
+      return (uint)qHash(idx.sibling(idx.row(), keyColumn).data(keyRole).toString(), seed);
     });
   }
